+++ src/onegov.onboarding/onegov/onboarding/app.py
@@ -1,8 +1,6 @@
-from hipchat import HipChat
 from onegov.core import Framework, utils
 from onegov.file import DepotApp
 from onegov.libres import LibresIntegration
-from onegov.onboarding import log
 from onegov.onboarding.theme import OnboardingTheme
 
 
@@ -14,21 +12,6 @@ class OnboardingApp(Framework, LibresIntegration, DepotApp):
         self.onboarding = cfg['onboarding']
         assert 'onegov.town' in self.onboarding
         assert 'namespace' in self.onboarding['onegov.town']
-
-    def notify_hipchat(self, message):
-        if 'hipchat' in self.onboarding:
-            try:
-                hipchat = HipChat(token=self.onboarding['hipchat']['token'])
-                hipchat.message_room(
-                    room_id=self.onboarding['hipchat']['room_id'],
-                    message_from='Onboarding',
-                    message=message,
-                    message_format='html',
-                    color='green',
-                    notify=True
-                )
-            except Exception:
-                log.exception("Error during Hipchat message")
 
 
 @OnboardingApp.static_directory()